pyproject.toml
[project]
name = "had-meeg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cebra>=0.5.0",
    "cupy-cuda11x>=13.6.0",
    "flask>=3.1.2",
    "graphviz>=0.21",
    "imageio>=2.37.2",
    "ipywidgets>=8.1.8",
    "kaleido>=1.2.0",
    "loguru>=0.7.3",
    "meegkit>=0.1.7",
    "mne>=1.6.1",
    "mne-bids>=0.14",
    "mne-icalabel>=0.8",
    "nbformat>=5.10.4",
    "nibabel>=5.3.2",
    "nilearn>=0.12.1",
    "onnxruntime>=1.23.2",
    "optuna>=4.6.0",
    "plotly>=6.5.0",
    "pycortex>=1.2.11",
    "pyctf-nihmegcore>=1.1.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "pyprep>=0.4.3",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "torchview>=0.2.7",
    "torchvision>=0.24.1",
    "tqdm-joblib>=0.0.5",
]
